book-store: use fixed point output (#1018)
* book-store: use fixed point output

Update generator for latest canonical-data.json
which uses integer expected values instead of floating point.

Update exercise to use type int for Cost output.
Add hints.md for guidance.
Update README.md.

+++ exercises/book-store/.meta/gen.go
@@ -34,7 +34,7 @@ type OneCase struct {
 	Input       struct {
 		Basket []int
 	}
-	Expected float64
+	Expected int
 }
 
 // template applied to above data structure generates the Go test cases
@@ -45,12 +45,12 @@ var tmpl = `package bookstore
 var testCases = []struct {
 	description string
 	basket      []int
-	expected    float64
+	expected    int
 }{{range .J.Groups}}{
 {{range .Cases}} {
 	description: "{{.Description}}",
 	basket:      {{printf "%#v" .Input.Basket}},
-	expected:    {{printf "%.2f" .Expected}},
+	expected:    {{printf "%d" .Expected}},
 },
 {{end}}}{{end}}
 `
